+++ Arrangement_on_surface_2/include/CGAL/Arr_geodesic_arc_on_sphere_traits_2.h
@@ -26,7 +26,7 @@
 
 #include <fstream>
 
-#include <boost/variant.hpp>
+#include <variant>
 
 #include <CGAL/config.h>
 #include <CGAL/Cartesian.h>
@@ -2063,7 +2063,7 @@ public:
      */
     template <typename OutputIterator>
     OutputIterator operator()(const Curve_2& c, OutputIterator oi) const {
-      typedef boost::variant<Point_2, X_monotone_curve_2>
+      typedef std::variant<Point_2, X_monotone_curve_2>
         Make_x_monotone_result;
       // std::cout << "full: " << c.is_full() << std::endl;
       // std::cout << "vert: " << c.is_vertical() << std::endl;
@@ -2321,8 +2321,7 @@ public:
                                         OutputIterator oi) const
     {
       typedef std::pair<Point_2, Multiplicity>          Intersection_point;
-      typedef boost::variant<Intersection_point, X_monotone_curve_2>
-                                                        Intersection_result;
+
       const Kernel& kernel = m_traits;
       typename Kernel::Equal_2 equal = kernel.equal_2_object();
 
@@ -2337,14 +2336,14 @@ public:
       if (equal(l1, r1)) {
         bool is_full = equal(l2, r2);
         X_monotone_curve_2 xc(l2_3, r2_3, normal, vertical, true, is_full);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
       if (equal(l2, r2)) {
         CGAL_assertion(! equal(l1, r1));        // already handled above
         X_monotone_curve_2 xc(l1_3, r1_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2358,19 +2357,19 @@ public:
         // 5. l1 = r2 < r1 < l2 = l1 | One overlap (handled above)
         if (in_between(r1, r2, l2)) {
           // Case 1.
-          *oi++ = Intersection_result(Intersection_point(l1_3, 1));
+          *oi++ = Intersection_point(l1_3, 1);
           return oi;
         }
         if (equal(r1, l2)) {
           // Case 2.
-          *oi++ = Intersection_result(Intersection_point(l1_3, 1));
-          *oi++ = Intersection_result(Intersection_point(l2_3, 1));
+          *oi++ = Intersection_point(l1_3, 1);
+          *oi++ = Intersection_point(l2_3, 1);
           return oi;
         }
         CGAL_assertion(in_between(r1, l2, r2));
         // Case 3.
         X_monotone_curve_2 xc(l2_3, r1_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2383,12 +2382,12 @@ public:
         // 5. l1 < l1 = r1 = l2 < r2 | One overlap (handled above)
         if (in_between(r2, r1, l1)) {
           // Case 1.
-          *oi++ = Intersection_result(Intersection_point(l2_3, 1));
+          *oi++ = Intersection_point(l2_3, 1);
           return oi;
         }
         // Case 3.
         X_monotone_curve_2 xc(l1_3, r2_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2400,13 +2399,13 @@ public:
         if (in_between(r1, l2, r2) || equal(r1, r2)) {
           // Cases 1 & 2
           X_monotone_curve_2 xc(l1_3, r1_3, normal, vertical, true);
-          *oi++ = Intersection_result(xc);
+          *oi++ = xc;
           return oi;
         }
         // Case 3
         CGAL_assertion(in_between(r2, l2, r1));
         X_monotone_curve_2 xc(l2_3, r2_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2418,13 +2417,13 @@ public:
         if (in_between(l1, r2, l2)) {
           // Cases 1
           X_monotone_curve_2 xc(l2_3, r2_3, normal, vertical, true);
-          *oi++ = Intersection_result(xc);
+          *oi++ = xc;
           return oi;
         }
         // Case 3
         CGAL_assertion(in_between(l1, l2, l2));
         X_monotone_curve_2 xc(l1_3, r1_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2442,12 +2441,12 @@ public:
         if (in_between(l2, r2, l1)) {
           // Case 2
           X_monotone_curve_2 xc(l1_3, r1_3, normal, vertical, true);
-          *oi++ = Intersection_result(xc);
+          *oi++ = xc;
           return oi;
         }
         // Case 3
         X_monotone_curve_2 xc(l2_3, r1_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
 
@@ -2457,12 +2456,12 @@ public:
       // Case 4
       if (in_between(l1, r1, l2)) {
         X_monotone_curve_2 xc(l2_3, r2_3, normal, vertical, true);
-        *oi++ = Intersection_result(xc);
+        *oi++ = xc;
         return oi;
       }
       // Case 5
       X_monotone_curve_2 xc(l1_3, r2_3, normal, vertical, true);
-      *oi++ = Intersection_result(xc);
+      *oi++ = xc;
       return oi;
     }
 
@@ -2560,8 +2559,6 @@ public:
       typedef typename Kernel::Equal_3                  Equal_3;
 
       typedef std::pair<Point_2, Multiplicity>          Intersection_point;
-      typedef boost::variant<Intersection_point, X_monotone_curve_2>
-                                                        Intersection_result;
 
       const Kernel& kernel = m_traits;
 
@@ -2584,9 +2581,9 @@ public:
               (res && (xc1.is_directed_right() != xc2.is_directed_right())))
           {
             if (xc1.left().is_min_boundary() && xc2.left().is_min_boundary())
-              *oi++ = Intersection_result(Intersection_point(xc1.left(), 1));
+              *oi++ = Intersection_point(xc1.left(), 1);
             if (xc1.right().is_max_boundary() && xc2.right().is_max_boundary())
-              *oi++ = Intersection_result(Intersection_point(xc1.right(), 1));
+              *oi++ = Intersection_point(xc1.right(), 1);
             return oi;
           }
 
@@ -2594,11 +2591,11 @@ public:
            * the other arc is completely overlapping.
            */
           if (xc1.left().is_min_boundary() && xc1.right().is_max_boundary()) {
-            *oi++ = Intersection_result(xc2);
+            *oi++ = xc2;
             return oi;
           }
           if (xc2.left().is_min_boundary() && xc2.right().is_max_boundary()) {
-            *oi++ = Intersection_result(xc1);
+            *oi++ = xc1;
             return oi;
           }
           /*! Find an endpoint that does not coincide with a pole, and project
@@ -2646,12 +2643,12 @@ public:
       // Observe that xc1 and xc2 may share two endpoints.
       Point_2 ed = m_traits.construct_point_2_object()(v.direction());
       if (is_in_between(ed, xc1) && is_in_between(ed, xc2))
-        *oi++ = Intersection_result(Intersection_point(ed, 1));
+        *oi++ = Intersection_point(ed, 1);
 
       Vector_3 vo(kernel.construct_opposite_vector_3_object()(v));
       Point_2 edo = m_traits.construct_point_2_object()(vo.direction());
       if (is_in_between(edo, xc1) && is_in_between(edo, xc2))
-        *oi++ = Intersection_result(Intersection_point(edo, 1));
+        *oi++ = Intersection_point(edo, 1);
 
       return oi;
     }