+++ Mesh_3/doc/Mesh_3/Concepts/IntersectionGeometricTraits_3.h
@@ -55,9 +55,9 @@ Function object that constructs the intersection
 between a 3D segment and a 3D triangle.
 Partial model of `::Kernel::Intersect_3`. Provides the operators:
 
-- `boost::optional< boost::variant< Point_3, Segment_3 > > operator()(Segment_3 seg, Triangle_3 tr)`
+- `std::optional< std::variant< Point_3, Segment_3 > > operator()(Segment_3 seg, Triangle_3 tr)`
 
-- `boost::optional< boost::variant< Point_3, Segment_3 > > operator()(Triangle_3 tr, Segment_3 seg)`
+- `std::optional< std::variant< Point_3, Segment_3 > > operator()(Triangle_3 tr, Segment_3 seg)`
 
 which computes the intersection between the triangle and the segment.
 */