+++ Mesh_3/doc/Mesh_3/Concepts/MeshDomain_3.h
@@ -140,7 +140,7 @@ A function object to query whether a point is in
 the input domain or not. In the positive case, it outputs the
 subdomain which includes the query point. Provides the operator:
 
-`boost::optional<Subdomain_index> operator()(Point_3 p)`
+`std::optional<Subdomain_index> operator()(Point_3 p)`
 */
 typedef unspecified_type Is_in_domain;
 
@@ -150,11 +150,11 @@ intersection queries between the surface patches of the domain and
 objects of type `Segment_3`, `Ray_3` or
 `Line_3`. Provides the operators:
 
-`boost::optional<Surface_patch_index> operator()(Segment_3 s)`
+`std::optional<Surface_patch_index> operator()(Segment_3 s)`
 
-`boost::optional<Surface_patch_index> operator()(Ray_3 r)`
+`std::optional<Surface_patch_index> operator()(Ray_3 r)`
 
-`boost::optional<Surface_patch_index> operator()(Line_3 l)`
+`std::optional<Surface_patch_index> operator()(Line_3 l)`
 
 The return type of the operators tell whether or not the query intersects a
 surface patch. In the positive case, it provides (through operator*()) the