+++ Polyline_simplification_2/include/CGAL/Polyline_simplification_2/Squared_distance_cost.h
@@ -50,7 +50,7 @@ public:
   /// is model of  `PolylineSimplificationVertexBase_2`.
 
     template<class CDT>
-    boost::optional<typename CDT::Geom_traits::FT>
+    std::optional<typename CDT::Geom_traits::FT>
     operator()(const Constrained_triangulation_plus_2<CDT>& pct
                , typename Constrained_triangulation_plus_2<CDT>::Vertices_in_constraint_iterator vicq)const
   {