+++ Surface_mesh_simplification/doc/Surface_mesh_simplification/CGAL/Surface_mesh_simplification/Policies/Edge_collapse/Edge_length_cost.h
@@ -36,8 +36,8 @@ public:
 
   The argument `placement` is unused.
   */
-  boost::optional<typename Edge_profile::FT> operator()(const Edge_profile& profile,
-                                                        const boost::optional<typename Edge_profile::Point>& placement) const;
+  std::optional<typename Edge_profile::FT> operator()(const Edge_profile& profile,
+                                                        const std::optional<typename Edge_profile::Point>& placement) const;
 
   /// @}