+++ Triangulation/include/CGAL/Regular_triangulation.h
@@ -129,12 +129,12 @@ private:
   // Wrapper
   struct Power_side_of_power_sphere_for_non_maximal_dim_d
   {
-    boost::optional<Flat_orientation_d>* fop;
+    std::optional<Flat_orientation_d>* fop;
     Construct_flat_orientation_d cfo;
     In_flat_power_side_of_power_sphere_d ifpt;
 
     Power_side_of_power_sphere_for_non_maximal_dim_d(
-      boost::optional<Flat_orientation_d>& x,
+      std::optional<Flat_orientation_d>& x,
       Construct_flat_orientation_d const&y,
       In_flat_power_side_of_power_sphere_d const&z)
     : fop(&x), cfo(y), ifpt(z) {}
@@ -144,7 +144,7 @@ private:
     {
       if(!*fop)
         *fop=cfo(a,b);
-      return ifpt(fop->get(),a,b,p);
+      return ifpt(fop->value(),a,b,p);
     }
   };